+++ src/Events.php
@@ -202,8 +202,8 @@ class Events extends Plugin
             $event->rules['events/event-types/new'] = 'events/event-types/edit';
             $event->rules['events/event-types/<eventTypeId:\d+>'] = 'events/event-types/edit';
 
-            $event->rules['events/settings'] = 'events/base/settings';
-            $event->rules['events/settings/<settingsNavItem:{handle}>'] = 'events/base/settings';
+            $event->rules['events/settings'] = 'events/settings/index';
+            $event->rules['events/settings/<settingsNavItem:{handle}>'] = 'events/settings/index';
         });
     }